Scripts and Data Files
Note - The scripts now contain languarge to "create database test1".

In order for the data inports to work, please make sure that the csv files are in directory "wamp/tmp". If not, then MySQL will not know where to look (and there will most likely be privledge issues as well).