SQL Script Update and User Dummy Data Update
OK. So I included comments on the sql script file. I also got rid of the auto increment for the tables since we are have the ID's in the dummy data files. Lastly I included the queries so the user can select category, manufacturer, model and part number (all based on previous selections)